Extract dispatcher into its own class as we might want to inherit from it
