+++ lib/statsd/instrument.rb
@@ -392,6 +392,8 @@ require "statsd/instrument/statsd_datagram_builder"
 require "statsd/instrument/dogstatsd_datagram_builder"
 require "statsd/instrument/null_sink"
 require "statsd/instrument/udp_sink"
+require "statsd/instrument/buffer"
+require "statsd/instrument/dispatcher"
 require "statsd/instrument/batched_udp_sink"
 require "statsd/instrument/capture_sink"
 require "statsd/instrument/log_sink"
